Refactor the Controllers Module and Test Module to allow generation of short links using custom domain name
On the front end side:

1) The short links are no longer generated by client-side JavaScript. It is provided by the server. Front-end (namely the shortLinkHref function) just creates the HTML using short link

2) For the `/analytics` route, the server sends a JSON payload (like before). But now, this payload has 'short_link` and `short_link_path` properties.
The former contains the entire short link, with the domain. The latter contains only the URL path (i.e., '/l/abcd').
This bifurcation helps in creating enable/disable requests by the front-end (see updateStatus function).

On the server side:

1) The analytics function on the links (controllers) module is modfied to send two separate properties in the JSON payload, `short_link` and `short_link_path`.

2) Also, the route handler for `\l\shorten:id` has been modified; the object returned by it contains two properties: shortLink and shortLinkPath
This was done to allow tests to rely on the shortLinkPath, while the front-end (see function shortLinkHref on the front-end) can rely on the entire shortLink (with custom domain) without having to construct it